Add validation check for multi-year filer to LoanValidator
Add checks to LoanValidator to check if company is a multi year filer. If not then do not use the balance at period start in calculations.
Added method, getIsMultipleYearFiler, to BaseValidator. Copied from and then removed from EmployeesValidator. This method was duplicated in other validators and removed to use the one in BaseValidator. Knock on effect other validators as required changes to their constructors or constructors added.
Added getter and setter for CompanyService in BaseValidator and replaced called to the getter in other validators that used CompanyService
Updated tests accordingly to cater for above changes.

BI-5144