Upgrade spring-boot to latest version
When moving to 2.5.5 it was discovered that the original
MongoClientOptions in MongoConfig.java was deprecated. The
docs suggested moving to mongoClientSettings.

Moved to MongoClientSettings and introduced 2 new variables
for handling the connection to MongoDB as without them, we were
getting connection refused as it was using the default server
settings.

2 new env vars added for both host and port of mongoDB database.
These are used in the MongoDbConnectionPoolConfig class and are called
for use in the new previously mentioned MongoClientSettings class.

Mnaually added hibernate as a dep on pom, as it is no longer included
with latest spring-boot as default.

Updated spring-boot on pom to latest 2.5.5