Fix issue where removing spring data uri causes auth to fail
It was discovered that by manually setting the host and port
instead of using the uri, causes issues in some of our environments
as they need credencials to connect to it.

After investigation it was clear that we needed to use the
spring data mongo uri variable and also set a connection string
in the mongo settings config file.